refactor(util): split the 760-line util.ts into domain modules (Q1)
Pure organizational refactor — no logic change. The monolithic util.ts is split
into cohesive modules behind the same @aws/util exports:

- core.ts          list/number + date primitives (addLists, getDailyDates,
                   isSameDay, getQuarter, getMostRecentValue*, ...)
- transactions.ts  DTO->domain transforms, stock grouping, transaction
                   aggregation, currencies, start date
- dividends.ts     per-quarter / TTM dividend calcs + updateDividends
- returns.ts       getReturn, getPortfolioValues, getYieldPerYear
- csv-import.ts    DEGIRO CSV parsing + translation
- yahoo-mapping.ts Yahoo response -> Ticker

util.ts is now an 8-line barrel re-exporting these, so every consumer
(@aws/util, portfolio.ts, util.spec.ts) is unchanged. The one previously-private
helper used across modules (getMostRecentValueAtIndex) is now exported from core.

Verified by the unchanged util test suite (43 tests) + full lint/test/build.
